+++ src/swgoh_comlink/StatCalc/__init__.py
@@ -1,3 +1,4 @@
 from .calculator import StatCalc
+from .calculator_async import StatCalcAsync
 
-__all__ = ["StatCalc"]
+__all__ = ["StatCalc", "StatCalcAsync"]