+++ src/swgoh_comlink/__init__.py
@@ -3,12 +3,12 @@ from __future__ import annotations
 
 import logging
 
-from swgoh_comlink.StatCalc import StatCalc
+from swgoh_comlink.StatCalc import StatCalc, StatCalcAsync
 from swgoh_comlink.swgoh_comlink import SwgohComlink
 from swgoh_comlink.swgoh_comlink_async import SwgohComlinkAsync
 from swgoh_comlink.version import __version__ as version
 
-__all__ = ["StatCalc", "SwgohComlink", "SwgohComlinkAsync", "version"]
+__all__ = ["StatCalc", "StatCalcAsync", "SwgohComlink", "SwgohComlinkAsync", "version"]
 
 # Follow Python best practice for library logging:
 # Only add NullHandler so the application controls all logging configuration.